+++ src/components/AnimateSharedLayout/utils/batcher.ts
@@ -1,3 +1,4 @@
+import { flushSync } from "framesync"
 import { VisualElement } from "../../../render/types"
 import { compareByDepth } from "../../../render/utils/compare-by-depth"
 import { Presence, SyncLayoutBatcher, SyncLayoutLifecycles } from "../types"
@@ -51,6 +52,15 @@ export function createBatcher(): SyncLayoutBatcher {
             order.forEach((child) => {
                 if (child.isPresent) child.presence = Presence.Present
             })
+
+            /**
+             * Starting these animations will have queued jobs on the frame loop. In some situations,
+             * like when removing an element, these will be processed too late after the DOM is manipulated,
+             * leaving a flash of incorrectly-projected content. By manually flushing these jobs
+             * we ensure there's no flash.
+             */
+            flushSync.preRender()
+            flushSync.render()
 
             queue.clear()
         },