+++ src/render/index.ts
@@ -4,7 +4,7 @@ import { Presence } from "../components/AnimateSharedLayout/types"
 import { Crossfader } from "../components/AnimateSharedLayout/utils/crossfader"
 import { MotionStyle } from "../motion/types"
 import { eachAxis } from "../utils/each-axis"
-import { copyAxisBox } from "../utils/geometry"
+import { axisBox, copyAxisBox } from "../utils/geometry"
 import {
     applyBoxTransforms,
     removeBoxTransforms,
@@ -23,17 +23,14 @@ import { variantPriorityOrder } from "./utils/animation-state"
 import { createLifecycles } from "./utils/lifecycles"
 import { updateMotionValuesFromProps } from "./utils/motion-values"
 import { updateLayoutDeltas } from "./utils/projection"
-import {
-    createLayoutState,
-    createProjectionState,
-    TargetProjection,
-} from "./utils/state"
+import { createLayoutState, createProjectionState } from "./utils/state"
 import { FlatTree } from "./utils/flat-tree"
 import {
     checkIfControllingVariants,
     checkIfVariantNode,
     isVariantLabel,
 } from "./utils/variants"
+import { Axis } from "../types/geometry"
 
 export const visualElement = <Instance, MutableState, Options>({
     treeType = "",
@@ -80,12 +77,12 @@ export const visualElement = <Instance, MutableState, Options>({
     const projection = createProjectionState()
 
     /**
-     * A reference to the nearest parent projection target. This is either
+     * A reference to the nearest projecting parent. This is either
      * undefined if we haven't looked for the nearest projecting parent,
      * false if there is no parent performing layout projection, or a reference
-     * to the parents prohection.
+     * to the projecting parent.
      */
-    let parentProjection: undefined | false | TargetProjection
+    let projectionParent: undefined | false | VisualElement
 
     /**
      * This is a reference to the visual state of the "lead" visual element.
@@ -700,7 +697,6 @@ export const visualElement = <Instance, MutableState, Options>({
          * Record the viewport box as it was before an expected mutation/re-render
          */
         snapshotViewportBox() {
-            // TODO: Store this snapshot in LayoutState
             element.prevViewportBox = element.measureViewportBox(false)
 
             /**
@@ -791,8 +787,19 @@ export const visualElement = <Instance, MutableState, Options>({
          * Update the projection of a single axis. Schedule an update to
          * the tree layout projection.
          */
-        setProjectionTargetAxis(axis, min, max) {
-            const target = projection.target[axis]
+        setProjectionTargetAxis(axis, min, max, isRelative = false) {
+            let target: Axis
+
+            if (isRelative) {
+                if (!projection.relativeTarget) {
+                    projection.relativeTarget = axisBox()
+                }
+                target = projection.relativeTarget[axis]
+            } else {
+                projection.relativeTarget = undefined
+                target = projection.target[axis]
+            }
+
             target.min = min
             target.max = max
 
@@ -812,6 +819,7 @@ export const visualElement = <Instance, MutableState, Options>({
             const { x, y } = element.getProjectionAnimationProgress()
 
             const shouldRebase =
+                !projection.relativeTarget &&
                 !projection.isTargetLocked &&
                 !x.isAnimating() &&
                 !y.isAnimating()
@@ -869,26 +877,33 @@ export const visualElement = <Instance, MutableState, Options>({
             sync.preRender(updateTreeLayoutProjection, false, true)
         },
 
-        resolveRelativeTargetBox() {
-            if (!projection.relativeTarget) return
+        getProjectionParent() {
+            if (projectionParent === undefined) {
+                let foundParent: VisualElement | false = false
 
-            if (parentProjection === undefined) {
-                let foundParentProjection: TargetProjection | false = false
+                // Search backwards through the tree path
+                for (let i = element.path.length - 1; i >= 0; i--) {
+                    const ancestor = element.path[i]
 
-                // Search backwards through tree
-                for (let i = element.path.length; i >= 0; i--) {
-                    const elementProjection = element.path[i].projection
-                    if (elementProjection.isEnabled) {
-                        foundParentProjection = elementProjection
+                    if (ancestor.projection.isEnabled) {
+                        foundParent = ancestor
+                        break
                     }
                 }
 
-                parentProjection = foundParentProjection
+                projectionParent = foundParent
             }
 
-            if (parentProjection) {
-                calcRelativeBox(projection, parentProjection)
-            } else {
+            return projectionParent
+        },
+
+        resolveRelativeTargetBox() {
+            if (!projection.relativeTarget) return
+
+            const relativeParent = element.getProjectionParent()
+
+            if (relativeParent) {
+                calcRelativeBox(projection, relativeParent.projection)
             }
         },