+++ src/render/types.ts
@@ -106,8 +106,14 @@ export interface VisualElement<Instance = any, RenderState = any>
     measureViewportBox(withTransform?: boolean): AxisBox2D
     updateLayoutMeasurement(): void
     getLayoutState: () => LayoutState
+    getProjectionParent: () => VisualElement | false
     getProjectionAnimationProgress(): MotionPoint
-    setProjectionTargetAxis(axis: "x" | "y", min: number, max: number): void
+    setProjectionTargetAxis(
+        axis: "x" | "y",
+        min: number,
+        max: number,
+        isRelative?: boolean
+    ): void
     startLayoutAnimation(axis: "x" | "y", transition: Transition): Promise<any>
     stopLayoutAnimation(): void
     snapshotViewportBox(): void